refactor: change app architecture
