+++ dnd-server/src/main/java/org/example/server/Controllers/ManualController.java
@@ -1,7 +1,7 @@
 package org.example.server.Controllers;
 
 import lombok.RequiredArgsConstructor;
-import org.example.server.Models.Room;
+import org.example.server.domain.Models.Room;
 import org.example.server.Services.MapService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
