+++ dnd-server/src/main/java/org/example/server/domain/Models/Cell.java
@@ -1,4 +1,4 @@
-package org.example.server.Models;
+package org.example.server.domain.Models;
 
 import lombok.*;
