+++ dnd-server/src/main/java/org/example/server/domain/Models/account/User.java
@@ -1,4 +1,4 @@
-package org.example.server.Models.account;
+package org.example.server.domain.Models.account;
 
 
 import jakarta.persistence.*;
