+++ dnd-server/src/main/java/org/example/server/domain/dto/RegistrationDto.java
@@ -1,4 +1,4 @@
-package org.example.server.dto;
+package org.example.server.domain.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
