+++ dnd-server/src/main/java/org/example/server/domain/dto/SignInRequest.java
@@ -1,4 +1,4 @@
-package org.example.server.dto;
+package org.example.server.domain.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
